pyproject.toml
[project]
name = "vyloc-backend"
version = "0.1.0"
description = "Vyloc - AI-Powered Product Localization Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.122.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "pillow>=11.0.0",
    "opencv-python>=4.10.0",
    "numpy>=2.0.0",
    "supabase>=2.10.0",
    "google-cloud-storage>=2.18.0",
    "aiofiles>=24.1.0",
    "httpx>=0.28.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "dodopayments>=0.18.0",
    "email-validator>=2.3.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "websockets>=12.0",
]

[project.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
